+++ source/apps/website/src/pages/CreateModel/components/ModelInfo/utils.tsx
@@ -10,17 +10,17 @@ export const getTrackTiles = (sortBy: SortByValue, filteringText: string) => {
 
   switch (sortBy) {
     case SortByValue.LENGTH_LONGEST:
-      sortedTracks = TRACKS.sort((a, b) => b.length - a.length);
+      sortedTracks = [...TRACKS].sort((a, b) => b.length - a.length);
       break;
     case SortByValue.DIFFICULTY_MOST:
-      sortedTracks = TRACKS.sort((a, b) => a.difficulty - b.difficulty);
+      sortedTracks = [...TRACKS].sort((a, b) => a.difficulty - b.difficulty);
       break;
     case SortByValue.DIFFICULTY_LEAST:
-      sortedTracks = TRACKS.sort((a, b) => b.difficulty - a.difficulty);
+      sortedTracks = [...TRACKS].sort((a, b) => b.difficulty - a.difficulty);
       break;
     case SortByValue.LENGTH_SHORTEST:
     default:
-      sortedTracks = TRACKS.sort((a, b) => a.length - b.length);
+      sortedTracks = [...TRACKS].sort((a, b) => a.length - b.length);
       break;
   }